(SERVER-2265) Refactor: move config utils
