+++ lib/puppetserver/ca/utils/config.rb
